+++ CHANGELOG.md
@@ -159,6 +159,7 @@
 - Allow displaying recent pickup history in profiles !1715 !1763 #989 @ChrisOelmueller
 - Given trust bananas will now notify the receiver via Bell !1795 #548 @muffelrudaijer
 - Topbar search query for numeric ID now returns profile link for that ID !1783 @ChrisOelmueller
+- Display same-day pickups when confirming to sign into a pickup slot !1827 @ChrisOelmueller
 
 ## Changes
 - Notify a user who is accepted to a workgroup with a bell #808 !1708 @alex.simm